ingestion/requirements.txt
psycopg2-binary
inflection